Tool registry stage 3: convert the 62 admin-tier tools to ToolDef domains
PR C-D of docs/TOOL-REGISTRY-DESIGN.md's strangler migration — every remaining
admin-tier closure tool (all of rbac.ts's ADMIN_TOOLS except generate_image,
which stage 1 already converted) moves out of buildToolServer into a
declarative domain file, bodies byte-verbatim, metadata mirrored from
rbac.ts/core.ts (cross-checked by tests/toolRegistry.test.ts), module state
moved exactly once:

- src/agent/tools/activity.ts: whats_new, user_history.
- src/agent/tools/moderation.ts: moderate, clear_warnings,
  list_member_warnings, list_muted_members, list_blocked_members,
  moderation_history + reserveWarnSlot/WARN_USER_RATE_LIMIT_PER_HOUR
  (applyManualWarnStrike already lives in tools/notify.ts — imported).
- src/agent/tools/appealsAdmin.ts: list_appeals, resolve_appeal.
- src/agent/tools/broadcast.ts: announce, create_poll, end_poll,
  create_thread, archive_thread (the latter four platforms:['discord'] per
  DISCORD_ONLY_TOOLS) + reserveAnnounceSlot/reservePollSlot/
  reservePollEndSlot/reserveThreadSlot and the ANNOUNCE/POLL/THREAD
  rate/bound constants.
- src/agent/tools/events.ts: create_event, cancel_event (both
  platforms:['discord']) + the EVENT_* bound constants.
- src/agent/tools/policyText.ts: set_community_guidelines,
  set_welcome_message + COMMUNITY_GUIDELINES_MAX_CHARS/
  WELCOME_MESSAGE_MAX_CHARS.
- src/agent/tools/knowledgeAdmin.ts: save_knowledge, list_knowledge,
  list_duplicate_knowledge, list_knowledge_conflicts, update_knowledge,
  delete_knowledge, merge_knowledge, list_knowledge_candidates,
  accept_knowledge_candidate, decline_knowledge_candidate,
  list_knowledge_gaps, list_answer_feedback, list_low_rated_knowledge,
  list_unhelpful_themes.
- src/agent/tools/accessAndSuggestions.ts: list_access_requests,
  list_suggestions, resolve_suggestion.
- src/agent/tools/roster.ts: add_member_note, list_member_notes,
  delete_member_note, list_roster, list_context_digests.
- src/agent/tools/digestsAdmin.ts: question_digest, admin_digest,
  review_queue, response_latency.
- src/agent/tools/reportsAdmin.ts: list_reports, resolve_report.
- src/agent/tools/membership.ts: add_member, remove_member, link_member,
  unlink_member + MEMBER_DM_FAILED_NOTE (ADMIN_DM_FAILED_NOTE stays with
  grant_admin in tools.ts).
- src/agent/tools/discordRoles.ts: assign_community_role,
  remove_community_role, list_assignable_roles (all platforms:['discord'])
  + checkAssignableRole as a local function now taking the turn's adapter
  explicitly (the one sanctioned body change: its two call sites pass
  `adapter` through).
- src/agent/tools/projectsAdmin.ts: project_create, project_add_member,
  project_remove_member, project_info, project_unbind_here, project_archive,
  project_unarchive, project_bind_here + the archivedSuffix local helper.
- tools.ts stays the barrel: every previously-exported symbol
  (WARN_USER_RATE_LIMIT_PER_HOUR, ANNOUNCE_RATE_LIMIT_PER_HOUR, the POLL_*/
  THREAD_*/EVENT_* bounds, COMMUNITY_GUIDELINES_MAX_CHARS,
  WELCOME_MESSAGE_MAX_CHARS) re-exported from its new home, so all existing
  import sites keep working. Only the 14 super-admin closure tools remain in
  buildToolServer; the server still registers
  [...registryTools, ...remaining closure tools] on one flat
  createSdkMcpServer (117 tools, unchanged — pinned by toolRegistry.test.ts).
- The share_project link-never-fetched SECURITY source scan in
  tests/tools.test.ts still read tools.ts and anchored on the now-removed
  `// --- Admin tools` banner (stage 2 moved share_project to social.ts but
  missed repointing this one) — now scans the WHOLE social.ts domain file,
  assertions unchanged, same module-boundary treatment the memberProjects.ts
  scan beside it already uses.
